Added auth for income routes

+++ frontend/src/components/Details.jsx
@@ -1,13 +1,22 @@
 import { useIncomeContext } from "../hooks/useIncomeContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 import Table from "react-bootstrap/Table";
 
 export default function Details(props) {
   const { dispatchIncome } = useIncomeContext();
+  const { user } = useAuthContext();
 
   const handleClick = async (id) => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("api/income/" + id, {
       method: "DELETE",
+      headers: {
+        "Authorization": `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
@@ -36,7 +45,14 @@ export default function Details(props) {
                 <td>{income.category}</td>
                 <td>{new Date(income.date).toLocaleDateString()}</td>
                 <td className="ps-0">
-                  <i className="bi bi-pencil me-2"></i>
+                  <i
+                    className="bi bi-pencil me-2"
+                    onClick={() =>
+                      console.log(
+                        new Date(income.date).toLocaleDateString("en-CA")
+                      )
+                    }
+                  ></i>
                   <i
                     className="bi bi-trash"
                     onClick={() => handleClick(income._id)}
